requirements.txt
future
emoji
requests
sqlalchemy
python-telegram-bot==11.1.0
psycopg2-binary
feedparser
aiohttp
aiofiles
ffmpeg-python
gTTS
dnspython
future
googletrans==4.0.0-rc1
motor
pillow
psutil
pykeyboard
pyrogram
python-arq
requests
search_engine_parser
speedtest-cli
TgCrypto
uvloop
youtube_dl
bs4
python-dotenv
feedparser
pyromod
fuzzysearch
img2pdf
telegraph
